requirements.txt
numpy==2.0.2
pygame==2.6.1
PyOpenGL==3.1.10
PyOpenGL-accelerate==3.1.10
